requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.6.2
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.0.2
django-appconf==1.0.6
django-author==1.2.0
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-content-editor==7.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-imagekit==5.0.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.4
django-phonenumber-field==8.0.0
django-rosetta==0.10.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
msgpack==1.1.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.43
phonenumberslite==8.13.46
pilkit==3.0
pillow==10.3.0
polib==1.2.0
prompt_toolkit==3.0.47
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
service-identity==24.1.0
setuptools==75.1.0
setuptools-git==1.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
zope.interface==7.0.3
